feat: single frame per error
Instead of storing the entire call-stack per error value we now only store a single frame per error value.

This works much better in my goasync library as the call-stack is of course not goroutine friendly but by saving a single frame from the location of the error wrapping, this helps generate a pseudo call-stack across goroutine boundaries.

Technically this is a backwards incompatible change but no one is really using this apart from myself and I can't be bothered with the whole v3 branch thing right now.